Move embeddings update to a function
