+++ site/gatsby-site/src/pages/cite/edit.js
@@ -11,14 +11,14 @@ import {
   FIND_REPORT_WITH_TRANSLATIONS,
   LINK_REPORTS_TO_INCIDENTS,
 } from '../../graphql/reports';
-import { UPDATE_INCIDENT, FIND_INCIDENTS } from '../../graphql/incidents';
+import { FIND_INCIDENTS } from '../../graphql/incidents';
 import { useMutation, useQuery } from '@apollo/client/react/hooks';
 import { format, getUnixTime } from 'date-fns';
 import { stripMarkdown } from '../../utils/typography';
 import { Formik } from 'formik';
 import pick from 'lodash/pick';
 import { useLocalization, LocalizedLink } from 'gatsby-theme-i18n';
-import { gql, useApolloClient } from '@apollo/client';
+import { gql } from '@apollo/client';
 import { useTranslation, Trans } from 'react-i18next';
 import RelatedIncidents from '../../components/RelatedIncidents';
 import { isEqual } from 'lodash';
@@ -55,35 +55,6 @@ const reportFields = [
   'embedding',
 ];
 
-const siblingsQuery = gql`
-  query Siblings($query: ReportQueryInput!) {
-    reports(query: $query) {
-      report_number
-      embedding {
-        from_text_hash
-        vector
-      }
-    }
-  }
-`;
-
-const incidentEmbedding = (reports) => {
-  reports = reports.filter((report) => report.embedding);
-  return reports.length == 0
-    ? null
-    : {
-        vector: reports
-          .map((report) => report.embedding.vector)
-          .reduce(
-            (sum, vector) => vector.map((component, i) => component + sum[i]),
-            Array(reports[0].embedding.vector.length).fill(0)
-          )
-          .map((component) => component / reports.length),
-
-        from_reports: reports.map((report) => report.report_number),
-      };
-};
-
 function EditCitePage(props) {
   const { t, i18n } = useTranslation();
 
@@ -118,8 +89,6 @@ function EditCitePage(props) {
   const addToast = useToastContext();
 
   const { config } = useLocalization();
-
-  const client = useApolloClient();
 
   const updateSuccessToast = ({ reportNumber, incidentId }) => ({
     message: (
@@ -212,76 +181,6 @@ function EditCitePage(props) {
               },
             },
           });
-
-          for (const incident of parentIncidents.incidents) {
-            // If the parent incident has changed,
-            // recompute its embedding not including that of the current report.
-            const exSiblingsResponse = await client.query({
-              query: siblingsQuery,
-              variables: {
-                query: {
-                  report_number_in: incident.reports
-                    .map((report) => report.report_number)
-                    .filter((report_number) => report_number != reportNumber),
-                },
-              },
-            });
-
-            const exSiblings = exSiblingsResponse.data.reports;
-
-            const embedding = incidentEmbedding(exSiblings);
-
-            if (embedding) {
-              await client.mutate({
-                mutation: UPDATE_INCIDENT,
-                variables: {
-                  query: { incident_id: incident.incident_id },
-                  set: { embedding },
-                },
-              });
-            }
-          }
-        }
-
-        for (const incident_id of values.incident_ids) {
-          // Update the embedding of the parent incident
-          // to include the current report.
-          const siblingIdsResponse = await client.query({
-            query: gql`
-              query SiblingIds($query: IncidentQueryInput) {
-                incident(query: $query) {
-                  incident_id
-                  reports {
-                    report_number
-                  }
-                }
-              }
-            `,
-            variables: { query: { incident_id } },
-          });
-
-          const siblingIds = siblingIdsResponse.data.incident.reports
-            .map((report) => report.report_number)
-            .filter((report_number) => report_number != reportNumber);
-
-          const siblingsResponse = await client.query({
-            query: siblingsQuery,
-            variables: { query: { report_number_in: siblingIds } },
-          });
-
-          const siblings = siblingsResponse.data.reports;
-
-          const embedding = incidentEmbedding(siblings.concat(values));
-
-          if (embedding) {
-            await client.mutate({
-              mutation: UPDATE_INCIDENT,
-              variables: {
-                query: { incident_id },
-                set: { embedding },
-              },
-            });
-          }
         }
 
         addToast(updateSuccessToast({ reportNumber, incidentId: values.incident_id }));
@@ -311,31 +210,6 @@ function EditCitePage(props) {
           },
         },
       });
-
-      for (const incident of parentIncidents.incidents) {
-        const exSiblingsResponse = await client.query({
-          query: siblingsQuery,
-          variables: {
-            query: {
-              report_number_in: incident.reports
-                .map((report) => report.report_number)
-                .filter((report_number) => report_number != reportNumber),
-            },
-          },
-        });
-
-        const exSiblings = exSiblingsResponse.data.reports;
-
-        const embedding = incidentEmbedding(exSiblings);
-
-        await client.mutate({
-          mutation: UPDATE_INCIDENT,
-          variables: {
-            query: { incident_id: incident.incident_id },
-            set: { embedding },
-          },
-        });
-      }
 
       addToast(deleteSuccessToast({ reportNumber }));
     } catch (e) {
